requirements.txt
arxiv>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
notion-client>=2.2.0
anthropic>=0.39.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.4.0
